+++ app/src/main/java/ng/com/thewhitecellfoundation/haemcam/ui/adapter/RegimenDrugAndCalendar.kt
@@ -2,12 +2,15 @@ package ng.com.thewhitecellfoundation.haemcam.ui.adapter
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.graphics.Color
 import android.util.AttributeSet
+import android.util.Log
 import android.view.LayoutInflater
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.airbnb.epoxy.CallbackProp
 import com.airbnb.epoxy.ModelProp
 import com.airbnb.epoxy.ModelView
+import com.applandeo.materialcalendarview.EventDay
 import ng.com.thewhitecellfoundation.haemcam.R
 import ng.com.thewhitecellfoundation.haemcam.databinding.RegimenDrugHorizontalBinding
 import ng.com.thewhitecellfoundation.haemcam.model.StringItemData
@@ -17,7 +20,7 @@ import java.util.*
     autoLayout = ModelView.Size.MATCH_WIDTH_WRAP_HEIGHT,
     defaultLayout = R.layout.regimen_drug_horizontal
 )
-class DrugNameView2 @JvmOverloads constructor(
+class RegimenDrugAndCalendar @JvmOverloads constructor(
     context: Context,
     attr: AttributeSet? = null,
     defStyleAttr: Int = 0
@@ -37,13 +40,25 @@ class DrugNameView2 @JvmOverloads constructor(
     @CallbackProp
     fun getDaysTimeData(listener: OnClickListener?) {
         if (listener != null) {
-            binding.calendarView.buildCalendar()
-            binding.regimenDrugDateTv.setOnClickListener(listener)
-            binding.calendarView.setDateFormat("yyyy MMM")
-            binding.calendarView.setPreventPreviousDate(true)
-            binding.calendarView.setErrToastMessage("You can not select the previous date.")
-            binding.calendarView.setOnDaySelectedListener { startDay, endDay ->
-                binding.regimenDrugDateTv.text = context.getString(R.string.days_cycle_formatted, startDay, endDay)
+            binding.calendarView.setOnDayClickListener {
+//                binding.regimenDrugDateTv.text = context.getString(R.string.days_cycle_formatted, startDay, endDay)
+                Log.i("RegimenDrugAndCalendar", it.calendar.timeInMillis.toString())
+                val startDate = it.calendar.timeInMillis
+                val calendars = ArrayList<Calendar>()
+                val midCalendars = ArrayList<Calendar>()
+                val events: MutableList<EventDay> = ArrayList()
+                var calendar: Calendar? = null
+                for (i in 1..3) {
+                    calendar = Calendar.getInstance()
+                    calendar.timeInMillis = startDate + (86400 * 1000 * i)
+                    calendars.add(calendar)
+                }
+
+                binding.calendarView.selectedDates = calendars
+                binding.calendarView.setHighlightedDays(calendars)
+                events.add(EventDay(calendar, null, Color.parseColor("#FF0000")))
+
+                binding.calendarView.setEvents(events)
             }
         }
     }
