+++ app/src/main/java/ng/com/thewhitecellfoundation/haemcam/ui/diagnosis/RegimenAndDrugsFragment.kt
@@ -3,22 +3,20 @@ package ng.com.thewhitecellfoundation.haemcam.ui.diagnosis
 import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import ng.com.thewhitecellfoundation.common.extensions.customOnDrawableRightListener
 import ng.com.thewhitecellfoundation.common.utils.viewBinding
 import ng.com.thewhitecellfoundation.haemcam.R
 import ng.com.thewhitecellfoundation.haemcam.databinding.ChemodrugBottomsheetLayoutBinding
-import ng.com.thewhitecellfoundation.haemcam.databinding.FragmentUserTreamentBinding
+import ng.com.thewhitecellfoundation.haemcam.databinding.FragmentRegimenAndDrugsBinding
 import ng.com.thewhitecellfoundation.haemcam.model.DataPair
-import ng.com.thewhitecellfoundation.haemcam.model.DrugDays
 import ng.com.thewhitecellfoundation.haemcam.model.OtherDrugDays
+import ng.com.thewhitecellfoundation.haemcam.model.Regimen
 import ng.com.thewhitecellfoundation.haemcam.model.StringItemData
-import ng.com.thewhitecellfoundation.haemcam.ui.adapter.drugDaysView
 import ng.com.thewhitecellfoundation.haemcam.ui.adapter.otherDrugDaysView
+import ng.com.thewhitecellfoundation.haemcam.ui.adapter.regimenAndDrugAdapter
 import ng.com.thewhitecellfoundation.haemcam.ui.home.ButtonAndProgressBarState
 import ng.com.thewhitecellfoundation.navigation.navigator.extensions.navigator
 import java.util.*
@@ -28,8 +26,8 @@ import java.util.*
  * Use the [UserTreamentFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class UserTreatmentFragment : Fragment(R.layout.fragment_user_treament) {
-    private val binding by viewBinding(FragmentUserTreamentBinding::bind)
+class RegimenAndDrugsFragment : Fragment(R.layout.fragment_regimen_and_drugs) {
+    private val binding by viewBinding(FragmentRegimenAndDrugsBinding::bind)
     lateinit var bottomSheetBinding: ChemodrugBottomsheetLayoutBinding
     lateinit var buttonAndProgressBarState: ButtonAndProgressBarState
 
@@ -85,31 +83,12 @@ class UserTreatmentFragment : Fragment(R.layout.fragment_user_treament) {
         }
 
         binding.regimenErcv.withModels {
-            DrugDays.listOfChemoTherapy.forEach { dd ->
-                Log.i("IDS", "${dd.id}")
-                drugDaysView {
+            Regimen.listOfChemoTherapy.forEach { dd ->
+                regimenAndDrugAdapter {
                     id(dd.id)
                     data(dd)
-                    binding.regimenTitleTv.customOnDrawableRightListener {
-                        val obj = DrugDays(
-                            R.string.regimen, R.array.regimen, null, R.string.chemo_drug,
-                            DataPair("", "")
-                        )
-                        Log.i("Object", "$obj")
-                        DrugDays.listOfChemoTherapy.sortBy { it.id }
-                        requestModelBuild()
-                    }
                     getDrugData(listOfRegimenDrugs)
                     getDaysTimeData { model, parentView, clickedView, position ->
-
-                        var currentTime = System.currentTimeMillis()
-                        val tz = TimeZone.getDefault()
-                        val cal = GregorianCalendar.getInstance(tz)
-                        val offsetInMillis = tz.getOffset(cal.timeInMillis)
-                        currentTime -= offsetInMillis.toLong()
-                        val date = Date(currentTime)
-                        Toast.makeText(context, "$date", Toast.LENGTH_SHORT).show()
-                        showDateTimeDialog(clickedView)
                     }
                     onDeleteListener { model, parentView, _, _ ->
                         model.data()?.remove
@@ -154,7 +133,7 @@ class UserTreatmentFragment : Fragment(R.layout.fragment_user_treament) {
     }
 
     companion object {
-        val firstRegimenChemo = DrugDays(
+        val firstRegimenChemo = Regimen(
             R.string.regimen, R.array.regimen, null, R.string.chemo_drug,
             DataPair("", "")
         )
