+++ app/src/main/res/layout/fragment_chemo_therapy_info.xml
@@ -53,7 +53,19 @@
             android:textSize="@dimen/_12sdp"
             android:text="@string/schedule_next_treatment"/>
 
-
+        <com.applandeo.materialcalendarview.CalendarView
+            android:id="@+id/calendar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toBottomOf="@id/schedule_tv"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:headerColor="@color/white"
+            app:headerLabelColor="@color/black"
+            app:todayLabelColor="@color/primaryVariant"
+            app:previousButtonSrc="@drawable/ic_keyboard_arrow_left_24"
+            app:forwardButtonSrc="@drawable/ic_keyboard_arrow_right_24"
+            />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.core.widget.NestedScrollView>
