+++ app/src/main/res/layout/fragment_other_drug_info.xml
@@ -53,8 +53,18 @@
             app:layout_constraintTop_toBottomOf="@id/chemodrug_ercv" />
 
         <com.applandeo.materialcalendarview.CalendarView
+            android:id="@+id/calendar"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"/>
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toBottomOf="@id/schedule_tv"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:headerColor="@color/white"
+            app:headerLabelColor="@color/black"
+            app:todayLabelColor="@color/primaryVariant"
+            app:previousButtonSrc="@drawable/ic_keyboard_arrow_left_24"
+            app:forwardButtonSrc="@drawable/ic_keyboard_arrow_right_24"
+            />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.core.widget.NestedScrollView>
