+++ app/src/main/res/layout/regimen_drug_horizontal.xml
@@ -32,7 +32,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/regimen_drug_name_tv" />
 
-    <pyxis.uzuki.live.sectioncalendarview.SectionCalendarView
+    <com.applandeo.materialcalendarview.CalendarView
         android:layout_width="match_parent"
         android:layout_height="@dimen/_240sdp"
         android:layout_marginTop="10dp"
@@ -40,6 +40,7 @@
         app:layout_constraintTop_toBottomOf="@id/regimen_drug_date_tv"
         android:animateLayoutChanges="true"
         android:visibility="gone"
+        app:type="many_days_picker"
         android:id="@+id/calendarView" />
 
 </androidx.constraintlayout.motion.widget.MotionLayout>
