+++ libraries.gradle
@@ -24,7 +24,7 @@ ext {
             'com.intuit.sdp:sdp-android:1.0.6',
             'com.google.firebase:firebase-analytics-ktx',
             //CalendarView
-            'com.github.WindSekirun:SectionCalendarView:1.0.5.1',
+//            'com.github.WindSekirun:SectionCalendarView:1.0.5.1',
             "org.naishadhparmar.zcustomcalendar:zcustomcalendar:1.0.1",
             'com.applandeo:material-calendar-view:1.7.0',
             //Spinner
